1er lot de refacto pour rendre le contenu hors carte côté serveur
