+++ app/voyage/page.tsx
@@ -1,17 +1,12 @@
-import dynamic from 'next/dynamic'
-// Map is forced as dynamic since it can't be rendered by nextjs server-side.
-// There is almost no interest to do that anyway, except image screenshots
-const Map = dynamic(() => import('./Map'), {
-	ssr: false,
-})
+// Server components here
+import Container from './Container'
 
-// But lots of content components that are not needing the map can be rendered
-// to provide link previews crucial for sharing place links
-
-const Page = ({ searchParams }) => (
-	<main id="voyage" style={{ height: '100%' }}>
-		<Map searchParams={searchParams} />
-	</main>
-)
+const Page = ({ searchParams }) => {
+	return (
+		<main id="voyage" style={{ height: '100%' }}>
+			<Container />
+		</main>
+	)
+}
 
 export default Page
